requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.4
django-allauth==65.3.1
docopt==0.6.2
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.6
num2words==0.5.14
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
razorpay==1.4.2
reportlab==4.2.5
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2024.2
XlsxWriter==3.2.2
whitenoise==6.6.0
cloudinary==1.32.0
django-cloudinary-storage==0.3.0
urllib3==1.26.18